# backend/openapi.yml
openapi: "3.0.3"
info:
  title: Progress Checker API
  version: "1.0.0"
paths:
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Authenticate user and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a mentor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/progress:
    get:
      operationId: listProgress
      summary: List latest progress by team
      security:
        - bearerAuth: []
      parameters:
        - name: team_id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Progress list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressListResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/teams/{teamId}/github-repos:
    post:
      operationId: registerRepository
      summary: Register a GitHub repository for a team
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRepoRequest'
      responses:
        '201':
          description: Repository registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: listRepositories
      summary: List GitHub repositories for a team
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Repository list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListReposResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/teams/{teamId}/github-repos/{repoId}:
    delete:
      operationId: removeRepository
      summary: Remove a GitHub repository
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Repository removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/teams/{teamId}/github-repos/{repoId}/token:
    put:
      operationId: updateToken
      summary: Update GitHub repository token
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTokenRequest'
      responses:
        '200':
          description: Token updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/staff/auth/login:
    post:
      operationId: staffLogin
      summary: Authenticate staff and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffLoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffLoginResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/staff/teams:
    get:
      operationId: staffListTeams
      summary: List all teams (staff only)
      security:
        - staffBearerAuth: []
      responses:
        '200':
          description: Team list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTeamsResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: staffCreateTeam
      summary: Create a new team (staff only)
      security:
        - staffBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTeamResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/staff/users:
    get:
      operationId: staffListUsers
      summary: List all users (staff only)
      security:
        - staffBearerAuth: []
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: staffCreateUser
      summary: Create a new user and generate setup URL (staff only)
      security:
        - staffBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/setup:
    post:
      operationId: setupPassword
      summary: Set password using setup token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupPasswordRequest'
      responses:
        '200':
          description: Password set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /internal/issues:
    post:
      operationId: createIssue
      summary: Create a GitHub issue (internal)
      security:
        - internalTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Issue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIssueResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    staffBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalTokenAuth:
      type: apiKey
      in: header
      name: X-Internal-Token

  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'

    UserResponse:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string

    ProgressListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TeamProgressResponse'

    TeamProgressResponse:
      type: object
      required: [team_id, team_name]
      properties:
        team_id:
          type: string
        team_name:
          type: string
        latest_progress:
          $ref: '#/components/schemas/LatestProgressResponse'

    LatestProgressResponse:
      type: object
      required: [id, participant_id, progress_bodies]
      properties:
        id:
          type: string
        participant_id:
          type: string
        progress_bodies:
          type: array
          items:
            $ref: '#/components/schemas/ProgressBodyResponse'

    ProgressBodyResponse:
      type: object
      required: [phase, sos, comment, submitted_at]
      properties:
        phase:
          type: string
        sos:
          type: boolean
        comment:
          type: string
        submitted_at:
          type: string
          format: date-time

    RegisterRepoRequest:
      type: object
      required: [github_repo_url, personal_access_token]
      properties:
        github_repo_url:
          type: string
        personal_access_token:
          type: string

    ListReposResponse:
      type: object
      required: [repos]
      properties:
        repos:
          type: array
          items:
            $ref: '#/components/schemas/RepoItem'

    RepoItem:
      type: object
      required: [id, owner, repo_name]
      properties:
        id:
          type: string
        owner:
          type: string
        repo_name:
          type: string

    UpdateTokenRequest:
      type: object
      required: [personal_access_token]
      properties:
        personal_access_token:
          type: string

    CreateIssueRequest:
      type: object
      required: [channel_id, title, body]
      properties:
        channel_id:
          type: string
        title:
          type: string
        body:
          type: string

    StaffLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    StaffLoginResponse:
      type: object
      required: [token, staff]
      properties:
        token:
          type: string
        staff:
          $ref: '#/components/schemas/StaffResponse'

    StaffResponse:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string

    CreateUserRequest:
      type: object
      required: [slack_user_id, name, email, role, team_id]
      properties:
        slack_user_id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [participant, mentor]
        team_id:
          type: string

    CreateUserResponse:
      type: object
      required: [user, setup_url]
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        setup_url:
          type: string

    ListUsersResponse:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'

    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string

    CreateTeamResponse:
      type: object
      required: [team]
      properties:
        team:
          $ref: '#/components/schemas/TeamResponse'

    TeamResponse:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string

    ListTeamsResponse:
      type: object
      required: [teams]
      properties:
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamResponse'

    SetupPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string

    CreateIssueResponse:
      type: object
      required: [issue_url]
      properties:
        issue_url:
          type: string